fix: fix linter errors
